backend/cli/requirements-cli.txt
click==8.1.7
requests==2.31.0
pyyaml==6.0.1
tabulate==0.9.0
rich==13.7.0
colorama==0.4.6
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-timeout==2.1.0
